Account: add, edit, delete profile
Refactoring account controller

+++ protected/controllers/AccountController.php
@@ -2,41 +2,50 @@
 
 class AccountController extends Controller
 {
-	public $content = '';
+	/**
+	 * @var User
+	 */
 	public $user;
+	public $breadcrumbs = ['Аккаунт' => ['/account']];
 
 	public function init()
 	{
-		$user_id = Yii::app()->user->id;
+		$user_id    = Yii::app()->user->id;
 		$this->user = User::model()->findByPk($user_id);
+		unset($this->user->pass);
+
+		$this->layout = 'account_layout';
 	}
 
+	/**
+	 * /account
+	 *
+	 * @throws CException
+	 */
 	public function actionProfile()
 	{
+		$this->pageTitle   = 'Личный кабинет';
+		$this->breadcrumbs = ['Аккаунт'];
 
-		$account = array(
+		$account = [
 			'values' => $this->user->attributes,
 			'labels' => $this->user->attributeLabels()
-		);
+		];
 		unset($account['values']['id'], $account['values']['pass']);
-
-		$this->pageTitle = 'Личный кабинет';
-		$this->content = $this->renderPartial('profile', array('account' => $account), TRUE);
-		$this->render('index', array('content' => $this->content, 'name' => $this->user->login));
+		$this->render('profile', ['account' => $account]);
 	}
 
+	/**
+	 * Page edit account: /account/edit
+	 */
 	public function actionEdit()
 	{
-		if (isset($_POST['User'])) {
-			$model = $this->user;
-			if (empty($_POST['User']['pass'])) {
-				$model->pass2 = $model->pass;
-				unset($_POST['User']['pass'], $_POST['User']['pass2']);
-			}
-			foreach ($_POST['User'] as $key => $value) {
-				$model->{$key} = $value;
-			}
-			if ($model->validate() && $model->save()) {
+		$this->pageTitle     = 'Редактирование аккаунта';
+		$this->breadcrumbs[] = 'Редактирование';
+
+		if ($attributes = Yii::app()->getRequest()->getPost(CHtml::modelName($this->user))) {
+			$this->user->setAttributes($attributes);
+			if ($this->user->validate() && $this->user->save()) {
 				Yii::app()->user->setFlash('success', "Аккаунт изменен");
 
 				$url = $this->createUrl('/account');
@@ -44,27 +53,39 @@ class AccountController extends Controller
 			}
 		}
 
-		$this->pageTitle = 'Редактирование аккаунта';
-		unset($this->user->pass);
-		$this->content = $this->renderPartial('edit', array('model' => $this->user), TRUE);
-		$this->render('index', array('content' => $this->content, 'name' => $this->user->login));
+		$this->render('edit', ['model' => $this->user]);
 	}
 
+	/**
+	 * My profiles: /account/passengers
+	 *
+	 * @throws CException
+	 */
 	public function actionPassengers()
 	{
-		$profiles = Profiles::model()
-							->findAllByAttributes(array('uid' => $this->user->id), array('order' => 'created DESC'));
+		$this->pageTitle     = 'Мои профили';
+		$this->breadcrumbs[] = 'Мои профили';
 
-		$this->pageTitle = 'Мои профили';
-		$this->content = $this->renderPartial('passengers', array('profiles' => $profiles), TRUE);
-		$this->render('index', array('content' => $this->content, 'name' => $this->user->login));
+		$profiles = Profiles::model()
+							->findAllByAttributes(['uid' => $this->user->id], ['order' => 'created DESC']);
+
+		$this->render('passengers', ['profiles' => $profiles]);
 	}
 
+	/**
+	 * Page add profile: /account/passengers/add
+	 *
+	 * @throws CException
+	 */
 	public function actionPassengersAdd()
 	{
+		$this->pageTitle                  = 'Создание нового профиля';
+		$this->breadcrumbs['Мои профили'] = ['/account/passengers'];
+		$this->breadcrumbs[]              = 'Добавление профиля';
+
 		$model = new Profiles();
-		if (isset($_POST['Profiles'])) {
-			$model->attributes = $_POST['Profiles'];
+		if ($attributes = Yii::app()->getRequest()->getPost(CHtml::modelName($model))) {
+			$model->setAttributes($attributes);
 			$model->uid = $this->user->id;
 			if ($model->validate() && $model->save()) {
 				Yii::app()->user->setFlash('success', "Профиль &laquo;" . $model->shortName() . "&raquo; создан");
@@ -73,44 +94,71 @@ class AccountController extends Controller
 				$this->redirect($url);
 			}
 		}
-		$this->pageTitle = 'Создание нового профиля';
-		$this->content = $this->renderPartial('one_passenger', array('model' => $model), TRUE);
-		$this->render('index', array('content' => $this->content, 'name' => $this->user->login));
+
+		$this->render('one_passenger', ['model' => $model]);
 	}
 
+	/**
+	 * @param $id
+	 *
+	 * @throws CDbException
+	 * @throws CHttpException
+	 */
 	public function actionPassengersDelete($id)
 	{
-		if ($model = Profiles::model()->findByPk($id)) {
-			if ($model->uid == $this->user->id) {
-				$model->delete();
+		$model = $this->loadProfile($id);
+		if ($model && $model->uid == $this->user->id) {
+			$model->delete();
 
-				Yii::app()->user->setFlash('success', "Профиль &laquo;" . $model->shortName() . "&raquo; удален");
+			Yii::app()->user->setFlash('success', "Профиль &laquo;" . $model->shortName() . "&raquo; удален");
+			$url = $this->createUrl('/account/passengers');
+			$this->redirect($url);
+		}
+		throw new CHttpException(404, "Страница не найдена");
+	}
+
+	/**
+	 * Edit profile page: /account/passengers/edit/<id>
+	 *
+	 * @param $id
+	 *
+	 * @return string
+	 * @throws CException
+	 * @throws CHttpException
+	 */
+	public function actionPassengersEdit($id)
+	{
+		$this->pageTitle                  = 'Редактирование профиля';
+		$this->breadcrumbs['Мои профили'] = ['/account/passengers'];
+		$this->breadcrumbs[]              = 'Редактирование профиля';
+
+		$model = $this->loadProfile($id);
+		if ($model && $model->uid == $this->user->id &&
+			$attributes = Yii::app()->getRequest()->getPost(CHtml::modelName($model))
+		) {
+			$model->setAttributes($attributes);
+			if ($model->validate() && $model->save()) {
+				Yii::app()->user->setFlash('success', "Профиль &laquo;" . $model->shortName() . "&raquo; обновлен");
 				$url = $this->createUrl('/account/passengers');
 				$this->redirect($url);
 			}
 		}
-		throw new CHttpException(400, "Страница не найдена");
+		$this->render('one_passenger', ['model' => $model]);
 	}
 
-	public function actionPassengersEdit($id)
+	/**
+	 * @param $id
+	 *
+	 * @return Profiles the loaded model
+	 * @throws CHttpException
+	 */
+	protected function loadProfile($id)
 	{
-		if ($model = Profiles::model()->findByPk($id)) {
-			if ($model->uid == $this->user->id) {
-				if (!empty($_POST['Profiles'])) {
-					$model->attributes = $_POST['Profiles'];
-					if ($model->validate() && $model->save()) {
-						Yii::app()->user->setFlash('success', "Профиль &laquo;" . $model->shortName() . "&raquo; обновлен");
-						$url = $this->createUrl('/account/passengers');
-						$this->redirect($url);
-					}
-				}
+		$model = Profiles::model()->findByPk($id);
+		if ($model === null)
+			throw new CHttpException(404, 'Страница не найдена');
 
-				$this->content = $this->renderPartial('one_passenger', array('model' => $model), TRUE);
-
-				return $this->render('index', array('content' => $this->content, 'name' => $this->user->login));
-			}
-		}
-		throw new CHttpException(400, "Страница не найдена");
+		return $model;
 	}
 
 	/**
